@@ -1,6 +1,6 @@
 source 'http://rubygems.org'
 
-gem 'rails', '3.1.1'
+gem 'rails', '5.2.0'
 
 # Bundle edge Rails instead:
 # gem 'rails',     :git => 'git://github.com/rails/rails.git'
@@ -13,22 +13,21 @@ gem 'rails', '3.1.1'
 #  gem 'activerecord-sqlserver-adapter'
 #  gem 'ruby-odbc'
 #end
-gem "pg", "~> 0.11.0"
+gem 'pg'
 
 # Gems used only for assets and not required
 # in production environments by default.
 group :assets do
-  gem 'sass-rails',   '~> 3.1.4'
-  gem 'coffee-rails', '~> 3.1.1'
-  gem 'uglifier', '>= 1.0.3'
+  gem 'sass-rails'
+  gem 'coffee-rails'
+  gem 'uglifier'
 end
 
 gem 'jquery-rails'
-gem "haml-rails", "~> 0.3.4"
-gem "haml", "~> 3.1.3"
-gem "inherited_resources"
-gem "stateflow"
-gem "formtastic", "~> 1.2.4"
+gem 'haml-rails'
+gem 'haml'
+gem 'stateflow'
+gem 'formtastic'
 
 # To use ActiveModel has_secure_password
 # gem 'bcrypt-ruby', '~> 3.0.0'
@@ -43,17 +42,18 @@ gem "formtastic", "~> 1.2.4"
 # gem 'ruby-debug19', :require => 'ruby-debug'
 
 group :development,:test do
-  gem 'simplecov', '>= 0.4.0', :require => false, :group => :test
+  gem 'simplecov', :require => false, :group => :test
   # Pretty printed test output
   gem 'turn', :require => false
-  gem "rspec-rails", "~> 2.7.0"
-  gem "factory_girl_rails", "~> 1.2.0"
-  gem "fuubar", "~> 0.0.5"
+  gem 'rspec-rails'
+  gem 'factory_bot_rails'
+  gem 'rails-controller-testing'
+  gem 'fuubar'
   #gem 'nokogiri', "~> 1.5.0"
   gem 'representative'
-  gem "guard-rspec"
+  gem 'guard-rspec'
   #gem 'guard-livereload'
-  gem "fakeweb"
-  gem "vcr"
-
+  gem 'webmock'
+  gem 'vcr'
+  gem 'pry'
 end
