@@ -1,4 +1,4 @@
-class AddAttributesToGame < ActiveRecord::Migration
+class AddAttributesToGame < ActiveRecord::Migration[5.0]
   def change
     add_column :games, :server_hits, :integer
     add_column :games, :client_hits, :integer
