@@ -1,4 +1,4 @@
-class CreateShips < ActiveRecord::Migration
+class CreateShips < ActiveRecord::Migration[5.0]
   def change
     create_table :ships do |t|
       t.string :name
