@@ -1,4 +1,4 @@
-class AddUserToGame < ActiveRecord::Migration
+class AddUserToGame < ActiveRecord::Migration[5.0]
   def change
     add_column :games, :email, :string
     add_column :games, :name, :string
