@@ -1,4 +1,4 @@
-class AddServerGameIdToGame < ActiveRecord::Migration
+class AddServerGameIdToGame < ActiveRecord::Migration[5.0]
   def change
     add_column :games, :server_game_id, :integer
   end
