@@ -1,4 +1,4 @@
-class AddServerNukeToNuke < ActiveRecord::Migration
+class AddServerNukeToNuke < ActiveRecord::Migration[5.0]
   def change
     add_column :nukes, :server_nuke, :boolean
   end
