@@ -1,4 +1,4 @@
-class RenameServerNukeInNuke < ActiveRecord::Migration
+class RenameServerNukeInNuke < ActiveRecord::Migration[5.0]
   def change
     rename_column :nukes, :server_nuke, :server_nuke_boolean
   end
