@@ -1,4 +1,4 @@
-class CreateBlocks < ActiveRecord::Migration
+class CreateBlocks < ActiveRecord::Migration[5.0]
   def change
     create_table :blocks do |t|
       t.integer :game_id
