@@ -1,4 +1,4 @@
-class CreateServerShips < ActiveRecord::Migration
+class CreateServerShips < ActiveRecord::Migration[5.0]
   def change
     create_table :server_ships do |t|
       t.string :name
