@@ -1,4 +1,4 @@
-class AddAttributesToServerShip < ActiveRecord::Migration
+class AddAttributesToServerShip < ActiveRecord::Migration[5.0]
   def change
     add_column :server_ships, :game_id, :integer
   end
