@@ -1,28 +1,5 @@
-require 'spec_helper'
-
-# This spec was generated by rspec-rails when you ran the scaffold generator.
-# It demonstrates how one might use RSpec to specify the controller code that
-# was generated by Rails when you ran the scaffold generator.
-#
-# It assumes that the implementation code is generated by the rails scaffold
-# generator.  If you are using any extension libraries to generate different
-# controller code, this generated spec may or may not pass.
-#
-# It only uses APIs available in rails and/or rspec-rails.  There are a number
-# of tools you can use to make these specs even more expressive, but we're
-# sticking to rails and rspec-rails APIs to keep things simple and stable.
-#
-# Compared to earlier versions of this generator, there is very limited use of
-# stubs and message expectations in this spec.  Stubs are only used when there
-# is no simpler way to get a handle on the object needed for the example.
-# Message expectations are only used when there is no simpler way to specify
-# that an instance is receiving a specific message.
-
+require 'rails_helper'
 describe BlocksController do
-
-  # This should return the minimal set of attributes required to create a valid
-  # Block. As you add validations to Block, be sure to
-  # update the return value of this method accordingly.
   def valid_attributes
     {:game_id => 1}
   end
@@ -31,30 +8,30 @@ describe BlocksController do
     it "assigns all blocks as @blocks" do
       block = Block.create! valid_attributes
       get :index
-      assigns(:blocks).should eq([block])
+      expect(assigns(:blocks)).to eq([block])
     end
   end
 
   describe "GET show" do
     it "assigns the requested block as @block" do
       block = Block.create! valid_attributes
-      get :show, :id => block.id
-      assigns(:block).should eq(block)
+      get :show, params: { id: block.id }
+      expect(assigns(:block)).to eq(block)
     end
   end
 
   describe "GET new" do
     it "assigns a new block as @block" do
       get :new
-      assigns(:block).should be_a_new(Block)
+      expect(assigns(:block)).to be_a_new(Block)
     end
   end
 
   describe "GET edit" do
     it "assigns the requested block as @block" do
       block = Block.create! valid_attributes
-      get :edit, :id => block.id
-      assigns(:block).should eq(block)
+      get :edit, params: { id: block.id }
+      expect(assigns(:block)).to eq(block)
     end
   end
 
@@ -62,31 +39,30 @@ describe BlocksController do
     describe "with valid params" do
       it "creates a new Block" do
         expect {
-          post :create, :block => valid_attributes
+          post :create, params: { block: valid_attributes }
         }.to change(Block, :count).by(1)
       end
 
       it "assigns a newly created block as @block" do
-        post :create, :block => valid_attributes
-        assigns(:block).should be_a(Block)
-        assigns(:block).should be_persisted
+        post :create, params: { block: valid_attributes }
+        expect(assigns(:block)).to be_a(Block)
+        expect(assigns(:block)).to be_persisted
       end
 
       it "redirects to the created block" do
-        post :create, :block => valid_attributes
-        response.should redirect_to(Block.last)
+        post :create, params: { block: valid_attributes }
+        expect(response).to redirect_to(Block.last)
       end
     end
 
     describe "with invalid params" do
       it "assigns a newly created but unsaved block as @block" do
         # Trigger the behavior that occurs when invalid params are submitted
-        Block.any_instance.stub(:save).and_return(false)
-        post :create, :block => {}
-        assigns(:block).should be_a_new(Block)
+        allow_any_instance_of(Block).to receive(:save).and_return(false)
+        # Block.any_instance.stub(:save).and_return(false)
+        post :create, params: { block: { name: "Hello"} }
+        expect(assigns(:block)).to be_a_new(Block)
       end
-
-
     end
   end
 
@@ -94,24 +70,20 @@ describe BlocksController do
     describe "with valid params" do
       it "updates the requested block" do
         block = Block.create! valid_attributes
-        # Assuming there are no other blocks in the database, this
-        # specifies that the Block created on the previous line
-        # receives the :update_attributes message with whatever params are
-        # submitted in the request.
-        Block.any_instance.should_receive(:update_attributes).with({'these' => 'params'})
-        put :update, :id => block.id, :block => {'these' => 'params'}
+        put :update, params: { id: block.id, block: {'x' => '1'} }
+        expect(block.reload.x).to eq(1)
       end
 
       it "assigns the requested block as @block" do
         block = Block.create! valid_attributes
-        put :update, :id => block.id, :block => valid_attributes
-        assigns(:block).should eq(block)
+        put :update, params: { id: block.id, block: valid_attributes }
+        expect(assigns(:block)).to eq(block)
       end
 
       it "redirects to the block" do
         block = Block.create! valid_attributes
-        put :update, :id => block.id, :block => valid_attributes
-        response.should redirect_to(block)
+        put :update, params: { id: block.id, block: valid_attributes }
+        expect(response).to redirect_to(block)
       end
     end
 
@@ -119,12 +91,11 @@ describe BlocksController do
       it "assigns the block as @block" do
         block = Block.create! valid_attributes
         # Trigger the behavior that occurs when invalid params are submitted
-        Block.any_instance.stub(:save).and_return(false)
-        put :update, :id => block.id, :block => {}
-        assigns(:block).should eq(block)
+        # Block.any_instance.stub(:save).and_return(false)
+        allow_any_instance_of(Block).to receive(:save).and_return(false)
+        put :update, params: { id: block.id, block: { name: "Hello"} }
+        expect(assigns(:block)).to eq(block)
       end
-
-
     end
   end
 
@@ -132,11 +103,8 @@ describe BlocksController do
     it "destroys the requested block" do
       block = Block.create! valid_attributes
       expect {
-        delete :destroy, :id => block.id
+        delete :destroy, params: { id: block.id }
       }.to change(Block, :count).by(-1)
     end
-
-
   end
-
 end
