@@ -1,60 +1,42 @@
-require 'spec_helper'
-
-# This spec was generated by rspec-rails when you ran the scaffold generator.
-# It demonstrates how one might use RSpec to specify the controller code that
-# was generated by Rails when you ran the scaffold generator.
-#
-# It assumes that the implementation code is generated by the rails scaffold
-# generator.  If you are using any extension libraries to generate different
-# controller code, this generated spec may or may not pass.
-#
-# It only uses APIs available in rails and/or rspec-rails.  There are a number
-# of tools you can use to make these specs even more expressive, but we're
-# sticking to rails and rspec-rails APIs to keep things simple and stable.
-#
-# Compared to earlier versions of this generator, there is very limited use of
-# stubs and message expectations in this spec.  Stubs are only used when there
-# is no simpler way to get a handle on the object needed for the example.
-# Message expectations are only used when there is no simpler way to specify
-# that an instance is receiving a specific message.
+require 'rails_helper'
 
 describe GamesController do
-
-  # This should return the minimal set of attributes required to create a valid
-  # Game. As you add validations to Game, be sure to
-  # update the return value of this method accordingly.
   def valid_attributes
     {:email => "rgkoek@yahoo.com", :name => "Gerhard"}
+  end
+  
+  def invalid_attributes
+    { ship_id: 1, block_id: 1}
   end
 
   describe "GET index" do
     it "assigns all games as @games" do
       game = Game.create! valid_attributes
       get :index
-      assigns(:games).should eq([game])
+      expect(assigns(:games)).to eq([game])
     end
   end
 
   describe "GET show" do
     it "assigns the requested game as @game" do
       game = Game.create! valid_attributes
-      get :show, :id => game.id
-      assigns(:game).should eq(game)
+      get :show, params: { :id => game.id }
+      expect(assigns(:game)).to eq(game)
     end
   end
 
   describe "GET new" do
     it "assigns a new game as @game" do
       get :new
-      assigns(:game).should be_a_new(Game)
+      expect(assigns(:game)).to be_a_new(Game)
     end
   end
 
   describe "GET edit" do
     it "assigns the requested game as @game" do
       game = Game.create! valid_attributes
-      get :edit, :id => game.id
-      assigns(:game).should eq(game)
+      get :edit, params: { :id => game.id }
+      expect(assigns(:game)).to eq(game)
     end
   end
 
@@ -62,35 +44,35 @@ describe GamesController do
     describe "with valid params" do
       it "creates a new Game" do
         expect {
-          post :create, :game => valid_attributes
+          post :create, params: {:game => valid_attributes }
         }.to change(Game, :count).by(1)
       end
 
       it "assigns a newly created game as @game" do
-        post :create, :game => valid_attributes
-        assigns(:game).should be_a(Game)
-        assigns(:game).should be_persisted
+        post :create, params: {:game => valid_attributes }
+        expect(assigns(:game)).to be_a(Game)
+        expect(assigns(:game)).to be_persisted
       end
 
       it "redirects to the created game" do
-        post :create, :game => valid_attributes
-        response.should redirect_to(game_path(Game.last))
+        post :create, params: {:game => valid_attributes }
+        expect(response).to redirect_to(game_path(Game.last))
       end
     end
 
     describe "with invalid params" do
       it "assigns a newly created but unsaved game as @game" do
         # Trigger the behavior that occurs when invalid params are submitted
-        Game.any_instance.stub(:save).and_return(false)
-        post :create, :game => {}
-        assigns(:game).should be_a_new(Game)
+        allow_any_instance_of(Game).to receive(:save).and_return(false)
+        post :create, params: { :game => invalid_attributes }
+        expect(assigns(:game)).to be_a_new(Game)
       end
 
       it "re-renders the 'new' template" do
         # Trigger the behavior that occurs when invalid params are submitted
-        Game.any_instance.stub(:save).and_return(false)
-        post :create, :game => {}
-        response.should render_template("new")
+        allow_any_instance_of(Game).to receive(:save).and_return(false)
+        post :create, params: { :game => invalid_attributes }
+        expect(response).to render_template("new")
       end
     end
   end
@@ -99,24 +81,20 @@ describe GamesController do
     describe "with valid params" do
       it "updates the requested game" do
         game = Game.create! valid_attributes
-        # Assuming there are no other games in the database, this
-        # specifies that the Game created on the previous line
-        # receives the :update_attributes message with whatever params are
-        # submitted in the request.
-        Game.any_instance.should_receive(:update_attributes).with({'these' => 'params'})
-        put :update, :id => game.id, :game => {'these' => 'params'}
+        put :update, params: {:id => game.id, :game => {'name' => 'params'}}
+        expect(game.reload.name).to eq('params')
       end
 
       it "assigns the requested game as @game" do
         game = Game.create! valid_attributes
-        put :update, :id => game.id, :game => valid_attributes
-        assigns(:game).should eq(game)
+        put :update, params: {:id => game.id, :game => valid_attributes}
+        expect(assigns(:game)).to eq(game)
       end
 
       it "redirects to the game" do
         game = Game.create! valid_attributes
-        put :update, :id => game.id, :game => valid_attributes
-        response.should redirect_to(games_path)
+        put :update, params: {:id => game.id, :game => valid_attributes}
+        expect(response).to redirect_to(games_path)
       end
     end
 
@@ -124,17 +102,17 @@ describe GamesController do
       it "assigns the game as @game" do
         game = Game.create! valid_attributes
         # Trigger the behavior that occurs when invalid params are submitted
-        Game.any_instance.stub(:save).and_return(false)
-        put :update, :id => game.id, :game => {}
-        assigns(:game).should eq(game)
+        allow_any_instance_of(Game).to receive(:save).and_return(false)
+        put :update, params: { id: game.id, game: { ship_id: "Hello"} }
+        expect(assigns(:game)).to eq(game)
       end
 
       it "re-renders the 'edit' template" do
         game = Game.create! valid_attributes
         # Trigger the behavior that occurs when invalid params are submitted
-        Game.any_instance.stub(:save).and_return(false)
-        put :update, :id => game.id, :game => {}
-        response.should render_template("edit")
+        allow_any_instance_of(Game).to receive(:save).and_return(false)
+        put :update, params: { id: game.id, game: { ship_id: "Hello"} }
+        expect(response).to render_template("edit")
       end
     end
   end
@@ -143,14 +121,14 @@ describe GamesController do
     it "destroys the requested game" do
       game = Game.create! valid_attributes
       expect {
-        delete :destroy, :id => game.id
+        delete :destroy, params: { :id => game.id }
       }.to change(Game, :count).by(-1)
     end
 
     it "redirects to the games list" do
       game = Game.create! valid_attributes
-      delete :destroy, :id => game.id
-      response.should redirect_to(games_url)
+      delete :destroy, params: { :id => game.id }
+      expect(response).to redirect_to(games_url)
     end
   end
 
@@ -159,9 +137,9 @@ describe GamesController do
     it "assigns the game as @game" do
         game = Game.create! valid_attributes
         # Trigger the behavior that occurs when invalid params are submitted
-        Game.any_instance.stub(:save).and_return(false)
-        get :battle, :id => game.id
-        assigns(:game).should eq(game)
+        allow_any_instance_of(Game).to receive(:save).and_return(false)
+        get :battle, params: {:id => game.id}
+        expect(assigns(:game)).to eq(game)
     end
   end
 
@@ -169,15 +147,12 @@ describe GamesController do
     it "should respond with json data" do
       game = Game.create! valid_attributes
       # Trigger the behavior that occurs when invalid params are submitted
-      Game.any_instance.stub(:save).and_return(false)
+      allow_any_instance_of(Game).to receive(:save).and_return(false)
       VCR.use_cassette('game_request', :record => :new_episodes) do
-        get :register, :id => game.id
-
+        get :register, params: {:id => game.id}
       end
-      assigns(:game).should eq(game)
-      response.should redirect_to(battle_game_path(game))
+      expect(assigns(:game)).to eq(game)
+      expect(response).to redirect_to(battle_game_path(game))
     end
   end
-
-
 end
