@@ -44,11 +44,7 @@ module Battleships
     end
 
     # Configure sensitive parameters which will be filtered from the log file.
-    config.filter_parameters += [:password, :password_confirmation]
-
-    # Enable IdentityMap for Active Record, to disable set to false or remove the line below.
-    config.active_record.identity_map = true
-    ActiveRecord::IdentityMap.enabled = true
+    config.filter_parameters += [:password, :password_confirmation]  
 
     # Enable the asset pipeline
     config.assets.enabled = true
